Refactored to use requestData throughout
Added converter.py and validation.py which are basically copies from the formal modules of the same name but with adaption removed and replaced with Mr Eby's Rule Dispatch.

Accessing the data method on the form now validates it automatically and raises an error if the form fails validation. Accessing the requestData method automatically converts the data to requestData form by iterating the form's schema and calling each widget's convert method

Also fixed the test cases with new api