Fix up formish to work with less forgiving templating engine.
Formish was sending a lot of None instances to the template. Mostly,
this was because the schema data -> request data process was returning
None instances, mostly in the widget pre rendering stage.

There were are a couple of other quirks, but nothing too major. We also
managed to remove a stickking plaster ;-) in one of the templates.

While "fixing" the pre render methods I managed to remove a number of
bits of code duplication where the widget subclass had overridden
methods but did exactly the same as the overridden version of the
method.

Anyway, the result is that formish now appears to work with straight
Mako and is a more resilient and "correct" library as a result (until
someone finds bits I missed and we're not using/testing ;-)).