Make success/failure callback more flexible and less surprising.
If only one of the success/failure callbacks was provided then the other
was silently ignored.

Typically, they will be used in a pair but the decision was that there
is no reason to enforce the pair. So now none, one or both of the
callbacks can be provided and will be called as expected.

I also flipped the pair of validate function around so that it doesn't
go validate() -> _validate_and_call() -> validate(). Just seems neater
that way and it's easier to read.

Oh, I even documented validate() fully while I was there. /me pats
himself on the head.