Simplify signature of CachedTempFilestore's put and get to only specify the cache tag and headers.
The CachedTempFilestore was neither general enough to be used for
stashing any file (it forced the content type and didn't return the
headers) nor specific to form file uploads (it didn't include an arg for
the filename).

The only thing the CachedTempFilestore actually cares about is the cache
tag and file data so I stripped the API back a little to use the headers
for everything other than the cache tag and to return the headers.