Rewrite CachedTempFilestore to use a configurable backend.
CachedTempFilestore subclessed FileSystemHeaderedFilestore and then
overrode get and put to do different things. In the case of put it
actually changed the method signature too.

CachedTempFilestore now has a FileSystemHeaderedFilestore backend (by
default). Definitely nicer from the OO point of view, but it also allows
the CachedTempFilestore or the backend to be rewritten depending on
what's wanted.

Writing a replacement for FileSystemHeaderedFilestore has potential to
be generally useful but, writing a replacement for CachedTempFilestore
would allow the store to take full advantage of any cacheability the
actual storage can provide, i.e. a HEAD request on some resource or a
fast SQL query to see if the cache tag is up to date.