Fix SelectChoice selected() comparison.
The template was assuming that the value was a [scalar] and was sending
in value[0]. In fact, the value could be something like
{'foo': [], 'bar': []}.

The value is now passed to selected() unchanged. It's up to that method
to do the correct comparison.

In reality, this really just makes the SelectChoice template usable by
other widgets.