Make the FileResource safe and provide a quickstart for simple use.
The FileResource used a default cache directory of 'cache'. If the same
directory was is used to store other temporary file (compiled templates,
etc) then there was a danger those could be accidentally accessible.

FileResource.__init__ now assumes nothing - it's up to the developer to
provide correct paths - but there is a FileResource.quickstart class
method to help set up a resource for simple/testing purposes.