Fix sequence new item template to support unicode.
The sequence "new item" template metadata was base64 encoded. However,
base64 is a byte string encoding scheme and the template is a text
string. Instead of base84 encoding it now does sort-of URL path
encoding, using JavaScript's standard decodeURIComponent to reconstruct
the template in the browser.

In addition I noticed that the structure templates were emitting
metadata for sequences ... which didn't seem right. There is no metadata
for structures so it's been removed from the templates and the
metadata() method has been moved from the collection to the sequence
class.

One nice feature of this change is that the jquery base64 library is not
needed by formish anymore. Woohoo :).

I'm fairly sure this is ok (tested the JS with Firefox/Linux,
Chromium/Linux, Opera/Linux, IE8/WindowsXP and IE6/WindowsXP).