Rename Form.error to Form.alert and include the alert in the form's rendered output.
Form.error was, until recently, emitted by the templates but was
semi-replaced with Form.errors.

The option of including general form errors in the Form.errors with a
None or '' key was considered but there's a relatively subtle difference
between errors that are the result of schema problems (required, not
long enough, username already registered, etc) and errors that are not
directly related to the data but are very much related to the form's
correct processing (service down, optimistic lock failure, etc).

So, there is now Form.errors for schema-related problems and Form.alert
for a general form processing problem.

Form.error has been deprecated (it wasn't actually used anymore anyway)
and, for now, an alias of Form.alert.

As part of this change I've written a RenderableProperty that, using the
magic of descriptors and proxy objects makes a property callable at
which point it renders itself using a template.