Don't share the Form's list of CSS classes between instances.
Commit 099fd98366c450afd25ba169c03c3d90ec5261fd looks like an attempt to
fix this but didn't actually work. It just stopped 'error' from getting
added *multiple* times (it still got added once) which was just a
symptom of the actual issue.

I've added tests this time. There's a test for the actual error as well
as a test for the different ways of configuring the class list.